LayoutTests/fast/xmlhttprequest/xmlhttprequest-access-self-as-blob.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<input type="file" id="singleFile" name="upfile" onchange="onFileChanged()" />
<div id="console"></div>
<script>
description("Tests that you cannot XHR to the current file as a file-origin Blob.");

window.jsTestIsAsync = true;

const sourcePath = "resources/xmlhttprequest-access-self-as-blob-real.html";
var fileUrl;

window.addEventListener('message', function(event) {
    testPassed('Recvied message');
    if (event.data === 'done')
        finishJSTest();
});

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.setAllowUniversalAccessFromFileURLs(false);

    var singleFileInput = document.getElementById("singleFile");
    dragFilesOntoInput(singleFileInput, [sourcePath]);
}

function onFileChanged() {
    testPassed("Drag event received.");
    var file = document.getElementById("singleFile").files[0];
    testPassed("Generated file-origin blob successfully.");

    var frameTarget = document.createElement('iframe');
    frameTarget.src = sourcePath;
    document.body.appendChild(frameTarget);

    setTimeout(function() {
        window.frames[0].postMessage( { BURL: file }, '*');
        testPassed('Sent Blob URL to frame.');
    }, 0);
}

function moveMouseToCenterOfElement(element)
{
    var centerX = element.offsetLeft + element.offsetWidth / 2;
    var centerY = element.offsetTop + element.offsetHeight / 2;
    eventSender.mouseMoveTo(centerX, centerY);
    testPassed("Moved to center of file input.");
}

function dragFilesOntoInput(input, files) {
    debug("Got files: " + files);
    eventSender.beginDragWithFiles(files);
    moveMouseToCenterOfElement(input);
    eventSender.mouseUp();
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
